fix: code review fixes
- move ReadFromConfig struct to file_source location
- update kubernetes and file sources accordingly

Tested:
- Local build
