+++ lib/file-source/Cargo.toml
@@ -14,6 +14,9 @@ winapi = { version = "0.3", features = ["winioctl"] }
 crc = "3.0.0"
 glob = "0.3.0"
 scan_fmt = "0.2.6"
+vector-config = { path = "../vector-config", default-features = false }
+vector-config-common = { path = "../vector-config-common", default-features = false }
+vector-config-macros = { path = "../vector-config-macros", default-features = false }
 
 [dependencies.bstr]
 version = "1.1"
